Re-add the Subscription model
A Subscription is used to link Products with Customers, essentially as 
an instance of a Product.  Quotas for a Product get associated with the 
Subscription and to the Subject of the Customer or shared group related 
to the Customer.
Fix cardinality of the overview diagram, add the Feature model, and 
update the explanatory text throughout.
Update the Quotas text to describe how quota Subjects are associated 
with create and update calls on participating Member Nodes.